pyproject.toml
[project]
name = "auditflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sse-starlette>=1.8.2",
    "asyncpg>=0.30.0",
    "groq>=0.33.0",
    "python-dotenv>=1.2.1",
]
